icg_website-main/requirements.txt
asgiref==3.8.1
boto3==1.38.19
botocore==1.38.19
dj-database-url==2.3.0
Django==5.2.1
django-phonenumber-field==8.1.0
django-storages==1.14.6
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
phonenumbers==9.0.5
pillow==11.2.1
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
whitenoise==6.9.0
